Rename TreeWeighter -> Weighter, get rid of original Weighter class
