+++ projects/injection/public/SIREN/injection/Weighter.h
@@ -2,69 +2,126 @@
 #ifndef SIREN_Weighter_H
 #define SIREN_Weighter_H
 
-#include <tuple>                  // for tuple
-#include <memory>                 // for shared_ptr
-#include <vector>                 // for vector
+#include <map>                                           // for map
+#include <tuple>
+#include <memory>                                        // for shared_ptr
+#include <vector>                                        // for vector
 
 #include <cereal/cereal.hpp>
+#include <cereal/access.hpp>
 #include <cereal/archives/json.hpp>
 #include <cereal/archives/binary.hpp>
 #include <cereal/types/vector.hpp>
+#include <cereal/types/array.hpp>
+#include <cereal/types/set.hpp>
+#include <cereal/types/map.hpp>
 #include <cereal/types/polymorphic.hpp>
 #include <cereal/types/base_class.hpp>
 #include <cereal/types/utility.hpp>
 
-namespace siren { namespace interactions { class InteractionCollection; } }
+#include "SIREN/dataclasses/InteractionTree.h"  // for InteractionT...
+#include "SIREN/dataclasses/Particle.h"         // for Particle
+
 namespace siren { namespace dataclasses { class InteractionRecord; } }
 namespace siren { namespace detector { class DetectorModel; } }
+namespace siren { namespace distributions { class PrimaryInjectionDistribution; } }
+namespace siren { namespace distributions { class SecondaryInjectionDistribution; } }
 namespace siren { namespace distributions { class WeightableDistribution; } }
 namespace siren { namespace injection { class Injector; } }
+namespace siren { namespace injection { class PrimaryInjectionProcess; } }
+namespace siren { namespace injection { class SecondaryInjectionProcess; } }
+namespace siren { namespace injection { class PhysicalProcess; } }
 namespace siren { namespace math { class Vector3D; } }
 
 namespace siren {
 namespace injection {
 
-class LeptonWeighter {
+// Class handling weight calculation for a single pair of injection and physical processes
+template<typename ProcessType>
+class ProcessWeighter {
 private:
+    std::shared_ptr<siren::injection::PhysicalProcess> phys_process;
+    std::shared_ptr<ProcessType> inj_process;
+    std::vector<std::shared_ptr<typename ProcessType::InjectionType>> unique_gen_distributions;
+    std::vector<std::shared_ptr<siren::distributions::WeightableDistribution>> unique_phys_distributions;
+    std::shared_ptr<siren::detector::DetectorModel> detector_model;
+    std::vector<std::shared_ptr<typename ProcessType::InjectionType>> const & GetInjectionDistributions();
+    void Initialize();
+    double normalization;
+public:
+    double InteractionProbability(std::tuple<siren::math::Vector3D, siren::math::Vector3D> const & bounds, siren::dataclasses::InteractionRecord const & record) const;
+    double NormalizedPositionProbability(std::tuple<siren::math::Vector3D, siren::math::Vector3D> const & bounds, siren::dataclasses::InteractionRecord const & record) const;
+    double PhysicalProbability(std::tuple<siren::math::Vector3D, siren::math::Vector3D> const & bounds, siren::dataclasses::InteractionRecord const & record) const;
+    double GenerationProbability(siren::dataclasses::InteractionTreeDatum const & datum) const;
+    double EventWeight(std::tuple<siren::math::Vector3D, siren::math::Vector3D> const & bounds, siren::dataclasses::InteractionTreeDatum const & datum) const;
+    ProcessWeighter(std::shared_ptr<siren::injection::PhysicalProcess> phys_process, std::shared_ptr<ProcessType> inj_process, std::shared_ptr<siren::detector::DetectorModel> detector_model);
+
+}; // ProcessWeighter
+
+typedef ProcessWeighter<siren::injection::PrimaryInjectionProcess> PrimaryProcessWeighter;
+typedef ProcessWeighter<siren::injection::SecondaryInjectionProcess> SecondaryProcessWeighter;
+
+// Parent class for calculating event weights
+// Assumes there is a unique secondary physical process for each particle type
+class Weighter {
+private:
+    // Supplied by constructor
     std::vector<std::shared_ptr<Injector>> injectors;
     std::shared_ptr<siren::detector::DetectorModel> detector_model;
-    //TODO Think about whether we want to pass a CrossSection collection, or a vector of cross sections
-    //TODO Think about what to do with multiple neutrino primary types. Do we want to support mutiple types across one InteractionCollection, across one Injector, across one LeptonWeighter?
-    std::shared_ptr<siren::interactions::InteractionCollection> interactions;
-    std::vector<std::shared_ptr<siren::distributions::WeightableDistribution>> physical_distributions;
+    std::shared_ptr<siren::injection::PhysicalProcess> primary_physical_process;
+    std::vector<std::shared_ptr<siren::injection::PhysicalProcess>> secondary_physical_processes;
 
-    std::vector<std::tuple<std::shared_ptr<siren::distributions::WeightableDistribution>, std::shared_ptr<siren::detector::DetectorModel>, std::shared_ptr<siren::interactions::InteractionCollection>>> unique_distributions;
-    std::vector<unsigned int> common_gen_idxs;
-    std::vector<unsigned int> common_phys_idxs;
-    std::vector<std::vector<unsigned int>> distinct_gen_idxs_by_injector;
-    std::vector<std::vector<unsigned int>> distinct_physical_idxs_by_injector;
-    std::vector<std::tuple<std::shared_ptr<siren::detector::DetectorModel>, std::shared_ptr<siren::interactions::InteractionCollection>>> unique_contexts;
-    std::vector<unsigned int> context_idx_by_injector;
-    double normalization;
-
-    bool user_supplied_position_distribution = false;
+    // Calculated upon initialization
+    std::vector<std::shared_ptr<PrimaryProcessWeighter>> primary_process_weighters;
+    std::vector<
+      std::map<
+        siren::dataclasses::ParticleType,
+        std::shared_ptr<SecondaryProcessWeighter>
+      >
+    > secondary_process_weighter_maps;
 
     void Initialize();
 public:
-    //TODO Think about the relationship between interaction probability and the positional distribution. Check that the math works out
-    //TODO Add versions of these functions that take precomputed intersections
-    double InteractionProbability(std::shared_ptr<Injector const> injector, siren::dataclasses::InteractionRecord const & record) const;
-    double InteractionProbability(std::tuple<siren::math::Vector3D, siren::math::Vector3D> bounds, siren::dataclasses::InteractionRecord const & record) const;
-    double UnnormalizedPositionProbability(std::shared_ptr<Injector const> injector, siren::dataclasses::InteractionRecord const & record) const;
-    double UnnormalizedPositionProbability(std::tuple<siren::math::Vector3D, siren::math::Vector3D> bounds, siren::dataclasses::InteractionRecord const & record) const;
-    double NormalizedPositionProbability(std::tuple<siren::math::Vector3D, siren::math::Vector3D> bounds, siren::dataclasses::InteractionRecord const & record) const;
-    //TODO Add a function to check that we have the right match up of variables between generator and physical distribution
-    //TODO Figure out a way to check that physical and generation probabilities match, and ignore those when weighting
-    LeptonWeighter(std::vector<std::shared_ptr<Injector>> injectors, std::shared_ptr<siren::detector::DetectorModel> detector_model, std::shared_ptr<siren::interactions::InteractionCollection> interactions, std::vector<std::shared_ptr<siren::distributions::WeightableDistribution>> physical_distributions);
-    double EventWeight(siren::dataclasses::InteractionRecord const & record) const;
-    double SimplifiedEventWeight(siren::dataclasses::InteractionRecord const & record) const;
-};
+    double EventWeight(siren::dataclasses::InteractionTree const & tree) const;
+    Weighter(std::vector<std::shared_ptr<Injector>> injectors, std::shared_ptr<siren::detector::DetectorModel> detector_model, std::shared_ptr<siren::injection::PhysicalProcess> primary_physical_process, std::vector<std::shared_ptr<siren::injection::PhysicalProcess>> secondary_physical_processes);
+    Weighter(std::vector<std::shared_ptr<Injector>> injectors, std::shared_ptr<siren::detector::DetectorModel> detector_model, std::shared_ptr<siren::injection::PhysicalProcess> primary_physical_process);
+    Weighter(std::vector<std::shared_ptr<Injector>> injectors, std::string filename);
+    void SaveWeighter(std::string const & filename) const;
+    void LoadWeighter(std::string const & filename);
+
+    template<typename Archive>
+    void save(Archive & archive, std::uint32_t const version) const {
+        if(version == 0) {
+            archive(::cereal::make_nvp("Injectors", injectors));
+            archive(::cereal::make_nvp("DetectorModel", detector_model));
+            archive(::cereal::make_nvp("PrimaryPhysicalProcess", primary_physical_process));
+            archive(::cereal::make_nvp("SecondaryPhysicalProcesses", secondary_physical_processes));
+        } else {
+            throw std::runtime_error("Weighter only supports version <= 0!");
+        }
+    }
+
+    template<typename Archive>
+    void load(Archive & archive, std::uint32_t const version) const {
+        if(version == 0) {
+            archive(::cereal::make_nvp("Injectors", injectors));
+            archive(::cereal::make_nvp("DetectorModel", detector_model));
+            archive(::cereal::make_nvp("PrimaryPhysicalProcess", primary_physical_process));
+            archive(::cereal::make_nvp("SecondaryPhysicalProcesses", secondary_physical_processes));
+        } else {
+            throw std::runtime_error("Weighter only supports version <= 0!");
+        }
+    }
+
+}; // Weighter
+
 
 } //namespace injection
 } //namespace siren
 
-CEREAL_CLASS_VERSION(siren::injection::LeptonWeighter, 0);
+#include "Weighter.tcc"
+
+CEREAL_CLASS_VERSION(siren::injection::Weighter, 0);
 
 
 #endif // SIREN_Weighter_H
-
